pyproject.toml
[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | _build
  | build
  | dist
)/
'''

[tool.poetry]
name = "the_baby"
version = "0.0.1"
description = "A baby-monitor setup using the tp-link tapo"
authors = ["Samuel Pedro"]
license = "LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pytapo = "^2.3"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
black = "^22.8.0"
pytest = "^7.1.3"
mypy = "^0.971"
flake8 = "^5.0.4"
pytest-xdist = "^2.5.0"
pip-chill = "^1.0.1"
pre-commit = "^2.20.0"
pytest-cov = "^3.0.0"
snakeviz = "^2.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
